Fix issue 1031, avoid the signed/unsigned warnings produced by the upcoming gcc 3.3.
* subversion/libsvn_subr/error.c (svn_strerror): Add an explicit cast.

* subversion/tests/libsvn_delta/delta-window-test.h (delta_window_print): Use
  a signed rather than an unsigned variable.

* subversion/tests/libsvn_delta/random-test.c (generate_random_file): Add
  an explicit cast.

* subversion/libsvn_ra_svn/marshal.c
  (writebuf_push, readbuf_drain): Use signed rather than unsigned variables.
  (writebuf_write, readbuf_read): Add an explicit cast.


git-svn-id: https://svn.apache.org/repos/asf/subversion/trunk@845832 13f79535-47bb-0310-9956-ffa450edef68

+++ subversion/tests/libsvn_delta/random-test.c
@@ -179,7 +179,7 @@ generate_random_file (apr_uint32_t maxlen,
       while (seqlen-- > 0)
         {
           const int ch = (random_bytes
-                          ? random_bytes[r % bytes_range]
+                          ? (unsigned)random_bytes[r % bytes_range]
                           : r % bytes_range);
           if (buf == end)
             {
